services/audit_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

pydantic==2.10.5
pydantic-settings==2.7.1

SQLAlchemy==2.0.36
asyncpg==0.30.0

httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
tldextract==5.1.2

aio-pika==9.5.4

scrapy==2.12.0
playwright==1.49.1
scrapy-playwright==0.0.41

google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1

pytest==8.3.3
pytest-asyncio==0.24.0